+++ app/src/features/projects/view/components/Partnerships.tsx
@@ -146,7 +146,7 @@ const Partnerships: React.FC<IPartnershipsProps> = (props) => {
             {indigenous_partnerships?.map((indigenousPartnership: string, index: number) => {
               return (
                 <Typography component="dd" variant="body1" key={index}>
-                  {indigenousPartnership}
+                  {codes?.first_nations?.find((item: any) => item.id === indigenousPartnership)?.name}
                 </Typography>
               );
             })}
