+++ app/src/hooks/api/useProjectApi.ts
@@ -99,7 +99,7 @@ const useProjectApi = (axios: AxiosInstance) => {
   ): Promise<string> => {
     const { data } = await axios.get(`/api/project/${projectId}/attachments/${attachmentId}/getSignedUrl`, {
       params: { attachmentType: attachmentType },
-      paramsSerializer: (params) => {
+      paramsSerializer: (params: any) => {
         return qs.stringify(params);
       }
     });
@@ -116,7 +116,12 @@ const useProjectApi = (axios: AxiosInstance) => {
   const getProjectsList = async (
     filterFieldData?: IProjectAdvancedFilterRequest
   ): Promise<IGetProjectsListResponse[]> => {
-    const { data } = await axios.post(`/api/projects`, filterFieldData || {});
+    const { data } = await axios.get(`/api/project/list`, {
+      params: filterFieldData,
+      paramsSerializer: (params: any) => {
+        return qs.stringify(params, { arrayFormat: 'repeat', filter: (_prefix: any, value: any) => value || undefined });
+      }
+    });
 
     return data;
   };
@@ -145,7 +150,7 @@ const useProjectApi = (axios: AxiosInstance) => {
   ): Promise<IGetProjectForUpdateResponse> => {
     const { data } = await axios.get(`api/project/${projectId}/update`, {
       params: { entity: entities },
-      paramsSerializer: (params) => {
+      paramsSerializer: (params: any) => {
         return qs.stringify(params);
       }
     });
@@ -173,7 +178,7 @@ const useProjectApi = (axios: AxiosInstance) => {
    * @return {*}  {Promise<ICreateProjectResponse>}
    */
   const createProject = async (project: ICreateProjectRequest): Promise<ICreateProjectResponse> => {
-    const { data } = await axios.post('/api/project', project);
+    const { data } = await axios.post('/api/project/create', project);
 
     return data;
   };
@@ -368,7 +373,7 @@ const useProjectApi = (axios: AxiosInstance) => {
   const getProjectReportMetadata = async (projectId: number, attachmentId: number): Promise<IGetReportMetaData> => {
     const { data } = await axios.get(`/api/project/${projectId}/attachments/${attachmentId}/metadata/get`, {
       params: {},
-      paramsSerializer: (params) => {
+      paramsSerializer: (params: any) => {
         return qs.stringify(params);
       }
     });
@@ -523,7 +528,7 @@ export const usePublicProjectApi = (axios: AxiosInstance) => {
   ): Promise<string> => {
     const { data } = await axios.get(`/api/public/project/${projectId}/attachments/${attachmentId}/getSignedUrl`, {
       params: { attachmentType: attachmentType },
-      paramsSerializer: (params) => {
+      paramsSerializer: (params: any) => {
         return qs.stringify(params);
       }
     });
@@ -545,7 +550,7 @@ export const usePublicProjectApi = (axios: AxiosInstance) => {
   ): Promise<IGetReportMetaData> => {
     const { data } = await axios.get(`/api/public/project/${projectId}/attachments/${attachmentId}/metadata/get`, {
       params: {},
-      paramsSerializer: (params) => {
+      paramsSerializer: (params: any) => {
         return qs.stringify(params);
       }
     });
